added functionality (addButton and savingState)
Add button with added functionality to write the current list to a text
file in External storage. The file name consists of date and time.
List state is saved across onDestroy, by saving the list to a text file
in internal storage in the onStop method.
All text files are 'tab' delimited.